+++ apps/c/volna/initGaussianLandslide.h
@@ -1,7 +1,7 @@
-void initGaussianLandslide( float *center, float *values, float *mesh_xmin, float *A, float *t, float *lx, float *ly, float *v) {
-  float x = center[0];
-  float y = center[1];
-  values[3] = (*mesh_xmin-x)*(x<0.0f)-5.0f*(x>=0.0f)+
-      *A*(*t<1.0f/(*v))*exp(-1.0f* *lx* *lx*(x+3.0f-*v**t)*(x+3.0f-*v**t)-*ly**ly*y*y)
-      +*A*(*t>=1.0f/(*v))*exp(-*lx*(x+3.0f-1.0f)**lx*(x+3.0f-1.0f)-*ly**ly*y*y);
+void initGaussianLandslide( double *center, double *values, double *mesh_xmin, double *A, double *t, double *lx, double *ly, double *v) {
+  double x = center[0];
+  double y = center[1];
+  values[3] = (*mesh_xmin-x)*(x<0.0)-5.0*(x>=0.0)+
+      *A*(*t<1.0/(*v))*exp(-1.0* *lx* *lx*(x+3.0-*v**t)*(x+3.0-*v**t)-*ly**ly*y*y)
+      +*A*(*t>=1.0/(*v))*exp(-*lx*(x+3.0-1.0)**lx*(x+3.0-1.0)-*ly**ly*y*y);
 }
